adding the  /var/lib/nova/instances/_base/ephemeral
There seems to be an issue where nova on reboot gets into a muddle when
trying to check if the files for a VM still exist, and somehow it ends
up checking for a file called just âephemeralârather that âephemeral_<size>_<fstype>.
(It gets worse as it then tries to re-get the ephemeral disk from
Glance and fails because at this stage it doesnt have the users credentials).

Just creating an empty file works around the issue.